tidy fix 11

+++ tasks/omp/kovalev_k_radix_sort_batcher_merge/src/sourse.cpp
@@ -155,7 +155,7 @@ bool kovalev_k_radix_sort_batcher_merge_omp::TestTaskOpenMP::PreProcessingImpl()
 }
 
 bool kovalev_k_radix_sort_batcher_merge_omp::TestTaskOpenMP::RunImpl() {
-  if (omp_get_max_threads() > 2 * n_input_) {
+  if (static_cast<unsigned int>(omp_get_max_threads()) > 2 * n_input_) {
     bool ret = RadixSigned(0, n_input_);
     memcpy(tmp_, mas_, sizeof(long long int) * n_input_);
     return ret;
